Get Census data for the Church of England at a parish, diocesan or national level.

Get Census data for the Church of England at a parish, diocesan or national level.

Usage

coe_census_stats(nomis_code, level, areas, relative = FALSE)

Arguments

nomis_code

The Nomis code of required data. Must be one of coe_datasets().

level

The geographical level at which data are required. Must be one of 'parish', 'diocese' or 'england'.

areas

A character vector of parish codes or diocese numbers. If specified, the returned tibble will contain only data for these areas. Cannot be used with level = 'england'. NA values will be removed.

relative

Logical. Should returned table contain absolute statitiscs (e.g. persons/households) or proportions for each area?

Value

A tibble::tibble() containing census data for the Church of England

Examples

coe_census_stats(nomis_code =  "TS001", level = "england")
#> # A tibble: 1 × 3
#>   population lives_in_a_household lives_in_a_communal_establishment
#>        <dbl>                <dbl>                             <dbl>
#> 1   56490048             55504302                            985746
coe_census_stats(nomis_code = "TS001", level = "diocese", areas = 1)
#> # A tibble: 1 × 5
#>   diocese_name diocese_number population lives_in_a_household
#>   <chr>        <chr>               <dbl>                <dbl>
#> 1 Bath & Wells 1                  979315              959638.
#> # ℹ 1 more variable: lives_in_a_communal_establishment <dbl>