View nomis codes of available data

View nomis codes of available data

Usage

coe_datasets(description = TRUE)

Arguments

description

Logical. If TRUE, (the default), a tibble is returned that contains nomis codes and corresponding dataset descriptions. If false, a character vector containing only nomis codes is returned.

Value

A tibble of nomis codes and dataset descriptions, or a character vector of nomis codes

Examples

coe_datasets()
#> # A tibble: 15 × 2
#>    nomis_code description                                                       
#>    <chr>      <chr>                                                             
#>  1 TS001      TS001 - Number of usual residents in households and communal esta…
#>  2 TS004      TS004 - Country of birth                                          
#>  3 TS007A     TS007A - Age by 5 year age bands                                  
#>  4 TS021      TS021 - Ethnic group                                              
#>  5 TS030      TS030 - Religion                                                  
#>  6 TS037      TS037 - General health                                            
#>  7 TS062      TS062 - Social Classification (NS-SeC)                            
#>  8 TS066      TS066 - Economic activity status                                  
#>  9 TS067      TS067 - Highest level of qualification                            
#> 10 TS011      TS011 - Households by deprivation dimensions                      
#> 11 TS025      TS025 - Household language                                        
#> 12 TS044      TS044 - Accommodation type                                        
#> 13 TS045      TS045 - Car or van availability                                   
#> 14 TS054      TS054 - Tenure                                                    
#> 15 TS003      TS003 - Household composition