Get a (list of) comparative table(s) of local, diocesan and national data for a given parish

Extracts parish, diocese and national level data for a given parish, then base::rbinds them into a single table. Iterates over nomis_codes to output a list of such tables

Usage

coe_parish_in_context(
  parish_code,
  nomis_codes = coe_nomis_codes(description = FALSE),
  relative = FALSE
)

Arguments

parish_code

A single parish code. Character.

nomis_codes

A character vector of nomis_codes.

relative

Logical. Should outputs be relative? Default is FALSE.

Value

A list of coe_parish_data tibbles

Examples

coe_parish_in_context(parish_code = "370047", nomis_codes = "TS001")
#> ✔ Reading from coe-parishes-2024.
#> ✔ Range ''parishes''.
#> ✔ Reading from coeparishdata_parish-data.
#> ✔ Range ''TS001''.
#> Warning: number of items to replace is not a multiple of replacement length
#> $TS001
#> [1] "parish" "parish" "nation"
#>