Get census data for CoE parishes

Get 2021 ONS census data for any number of Church of England parishes

parish_stats(nomis_code, parish_codes = NULL, relative = FALSE)

Arguments

nomis_code

The nomis code for the data required must be a length-one character.

parish_codes

Character vector of parishes for which data are desired. If NULL (the default), all parishes are returned

relative

Logical. Should a tibble of relative values be returned? The default is absolute values.

Value

A single tibble

Examples

parish_stats("TS001") |> 
  head()
#> # A tibble: 6 × 4
#>   parish_code population lives_in_a_household lives_in_a_communal_establishment
#>   <chr>            <dbl>                <dbl>                             <dbl>
#> 1 100001              61                 50.2                             10.8 
#> 2 100004            9815               9767.                              48.4 
#> 3 100006            9656               9626.                              29.8 
#> 4 100007            3582               3505.                              76.9 
#> 5 100009             121                112.                               9.09
#> 6 100010           13239              12477.                             762.  

parish_stats("TS004", c("100001", "100006"))
#> # A tibble: 2 × 10
#>   parish_code population country_of_birth_united_kingdom country_of_birth_euro…¹
#>   <chr>            <dbl>                           <dbl>                   <dbl>
#> 1 100001              61                            50.9                    3.52
#> 2 100006            9656                          9013.                   290.  
#> # ℹ abbreviated name: ¹​country_of_birth_europe_eu_countries
#> # ℹ 6 more variables: country_of_birth_europe_non_eu_countries <dbl>,
#> #   country_of_birth_africa <dbl>, country_of_birth_middle_east_and_asia <dbl>,
#> #   country_of_birth_the_americas_and_the_caribbean <dbl>,
#> #   country_of_birth_antarctica_and_oceania_including_australasia_and_other <dbl>,
#> #   country_of_birth_british_overseas <dbl>